+++ app/src/main/java/fr/antoinehory/lejeudu5000/ui/feature_info/InfoUiState.kt
@@ -1,4 +1,0 @@
-package fr.antoinehory.lejeudu5000.ui.feature_info
-
-class InfoUiState {
-}